refactor: break up utils file
This serves two purposes

1) Ensuring this file doesn't get too large
2) Breaking out some of the utils that require external resources into
   their own file. Unfortunately, anything that imports 'vscode' cannot
   be run under mocha, so we can't import modules that import 'vscode'
   from a unit test. This was the _real_ reason for the change, but the
   organization is a nice bonus :)
