+++ package.json
@@ -2090,7 +2090,7 @@
         "format": "pre-commit run --all prettier && pre-commit run --all stylua-github",
         "keybind": "node ./scripts/keybindings/index.cjs",
         "test": "npm run test:unit && npm run test:integration",
-        "test:unit": "mocha ./out/test/unit",
+        "test:unit": "mocha ./out/test/unit/**/*.js",
         "test:integration": "node ./out/test/runIntegrationTest.js",
         "pretest:unit": "npm run test-compile",
         "pretest:integration": "npm run test-compile",
